TOOLS: move `CommandArg` to `tools-repl`
