+++ cli/src/main/java/org/aya/cli/repl/Repl.java
@@ -9,7 +9,6 @@ import org.aya.api.error.Problem;
 import org.aya.api.util.InterruptException;
 import org.aya.api.util.NormalizeMode;
 import org.aya.cli.repl.command.Command;
-import org.aya.cli.repl.command.CommandArg;
 import org.aya.cli.repl.command.CommandManager;
 import org.aya.cli.repl.jline.JlineRepl;
 import org.aya.cli.repl.jline.completer.AyaCompleters;
@@ -17,6 +16,7 @@ import org.aya.cli.single.CliReporter;
 import org.aya.cli.utils.MainArgs;
 import org.aya.prelude.GeneratedVersion;
 import org.aya.pretty.doc.Doc;
+import org.aya.repl.CommandArg;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jline.builtins.Completers;
@@ -42,13 +42,13 @@ public abstract class Repl implements Closeable, Runnable {
   }
 
   public CommandManager makeCommand() {
-    return new CommandManager(ImmutableSeq.of(
+    return new CommandManager(Repl.class, ImmutableSeq.of(
       CommandArg.STRING,
       CommandArg.STRICT_BOOLEAN,
       CommandArg.STRICT_INT,
       CommandArg.shellLike(Path.class, new Completers.FileNameCompleter(), this::resolveFile),
       CommandArg.from(ReplCommands.Code.class, new AyaCompleters.Code(this), ReplCommands.Code::new),
-      CommandArg.from(ReplCommands.HelpItem.class, new AyaCompleters.Help(commandManager), ReplCommands.HelpItem::new),
+      CommandArg.from(ReplCommands.HelpItem.class, new AyaCompleters.Help(() -> commandManager), ReplCommands.HelpItem::new),
       CommandArg.fromEnum(DistillerOptions.Key.class),
       CommandArg.fromEnum(NormalizeMode.class)
     ), ImmutableSeq.of(
