+++ cli/src/main/java/org/aya/cli/repl/command/CommandManager.java
@@ -6,7 +6,7 @@ import kala.collection.immutable.ImmutableMap;
 import kala.collection.immutable.ImmutableSeq;
 import kala.control.Try;
 import kala.tuple.Tuple;
-import org.aya.cli.repl.Repl;
+import org.aya.repl.CommandArg;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -20,7 +20,7 @@ public class CommandManager {
     @NotNull MethodHandle entry,
     @Nullable CommandArg argFactory
   ) {
-    public Command.Result invoke(@NotNull Repl repl, @NotNull String argument) throws Throwable {
+    public Command.Result invoke(@NotNull Object repl, @NotNull String argument) throws Throwable {
       if (argFactory != null) return (Command.Result) entry.invoke(owner, repl,
         Try.of(() -> argFactory.parse(argument)).getOrNull());
       else return (Command.Result) entry.invoke(owner, repl);
@@ -29,8 +29,14 @@ public class CommandManager {
 
   public final @NotNull ImmutableSeq<@NotNull CommandGen> cmd;
   public final @NotNull ImmutableMap<Class<?>, CommandArg> argFactory;
+  private final @NotNull Class<?> replClass;
 
-  public CommandManager(@NotNull ImmutableSeq<CommandArg> argFactory, @NotNull ImmutableSeq<Command> commands) {
+  public CommandManager(
+    @NotNull Class<?> replClass,
+    @NotNull ImmutableSeq<CommandArg> argFactory,
+    @NotNull ImmutableSeq<Command> commands
+  ) {
+    this.replClass = replClass;
     this.argFactory = argFactory.view().map(c -> Tuple.of(c.type(), c)).toImmutableMap();
     this.cmd = commands.map(this::genCommand);
   }
@@ -47,7 +53,7 @@ public class CommandManager {
       var param = handle.type().parameterList();
       if (param.size() < 2)
         throw new IllegalArgumentException("entry method must at least have 1 parameters (the `REPL` instance)");
-      if (param.get(1) != Repl.class)
+      if (param.get(1) != replClass)
         throw new IllegalArgumentException("entry method must have `Repl` as first parameter");
       if (param.size() == 2) return new CommandGen(c, handle, null);
       // TODO: support more than 1 parameter
@@ -64,7 +70,7 @@ public class CommandManager {
     @NotNull ImmutableSeq<CommandGen> command,
     @NotNull String argument
   ) {
-    public Command.Result run(@NotNull Repl repl) throws Throwable {
+    public Command.Result run(@NotNull Object repl) throws Throwable {
       return switch (command.size()) {
         case 1 -> command.first().invoke(repl, argument);
         case 0 -> Command.Result.err("Command `" + name + "` not found", true);
