+++ tools-repl/src/main/java/org/aya/repl/CommandArg.java
@@ -1,9 +1,8 @@
 // Copyright (c) 2020-2021 Yinsen (Tesla) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli.repl.command;
+package org.aya.repl;
 
 import kala.control.Try;
-import org.aya.repl.Completers;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jline.reader.Completer;
